requirements.txt
fastapi
uvicorn
pyyaml
python-dotenv
anthropic
celery
redis
streamlit
pytest
